requirements.txt
# Django
Django==3.2

# Rest Framework
requests==2.23.0
djangorestframework==3.12.0
django-oauth-toolkit==1.3.2
drf-yasg==1.20.0

